Improves language select loading state
Enhances the language select component by displaying a loading indicator within the select trigger while data is being fetched, improving the user experience.

Removes the skeleton component and introduces a `loading` prop to the `SelectTrigger` component. This prop conditionally renders a spinner icon when `isLoading` is true, providing visual feedback to the user that the language options are being loaded.